requirements.txt
asgiref==3.4.1
autopep8==1.5.7
boto3==1.18.21
botocore==1.21.21
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.45.0
django-multiselectfield==0.1.12
django-storages==1.11.1
django-taggit==1.5.1
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
gunicorn==20.1.0
haversine==2.3.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
openpyxl~=3.0.7